refactor(clerk-js,shared): Move `isDevOrStagingUrl` from clerk-js to shared
A new feature will need this logic available for more than one package. It is time to make this available for all our packages and eliminate duplication
